+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Option to hide the "Screen captured" text
 
+### 🐛 Fixed
+
+- Class Resources: Hide toggles now respect UF player frame activation when switching.
+
 ---
 
 ## [7.2.0] - 2026-01-21